requirements.txt
paho-mqtt==1.5.0
python-periphery==2.1.0
